refactor: extract LLM model, prompt compilation, and session logic into modular sub-services
